3697 add spatial resample (#3701)
* temp spatial_resample


* fixes resampling


* fixes precisions


* update dict version


* fixes unit tests


* adds docs


* copy grid for resampling


* fixes unit tests


* remove normalize coordinates


* [MONAI] python code formatting


* try to fix #3621 (#3673)


* fixes typing


* fixes grid_sample, interpolate URLs


* simplify norm_coords


* update docstring


* update moveaxis


* spatial sample tests


* additional tests spatial resample


* test invert saptial resample


* fixes unit tests


* rtol assert close


* fixes TF32 tests


* smaller tests


* skip when quick testing


* comp tensor and ndarray


* update based on comments


* fixes tests


* try to use torch.solve


* Revert "fixes tests"

This reverts commit e5324903341dde07ac65b8eb9d9e3c09e840c11d.


* update based on comments


* fixes test_affined


* default to float32 rotate/randrotate


* workaround for #3752


* default to float32 rotate/randrotate


* temp test


* update docstring

+++ tests/test_inverse.py
@@ -209,10 +209,15 @@ TESTS.append(("Zoomd 3d", "3D", 3e-2, Zoomd(KEYS, zoom=[2.5, 1, 3], keep_size=Fa
 
 TESTS.append(("RandZoom 3d", "3D", 9e-2, RandZoomd(KEYS, 1, [0.5, 0.6, 0.9], [1.1, 1, 1.05], keep_size=True)))
 
-TESTS.append(("RandRotated, prob 0", "2D", 0, RandRotated(KEYS, prob=0)))
+TESTS.append(("RandRotated, prob 0", "2D", 0, RandRotated(KEYS, prob=0, dtype=np.float64)))
 
 TESTS.append(
-    ("Rotated 2d", "2D", 8e-2, Rotated(KEYS, random.uniform(np.pi / 6, np.pi), keep_size=True, align_corners=False))
+    (
+        "Rotated 2d",
+        "2D",
+        8e-2,
+        Rotated(KEYS, random.uniform(np.pi / 6, np.pi), keep_size=True, align_corners=False, dtype=np.float64),
+    )
 )
 
 TESTS.append(
@@ -220,7 +225,7 @@ TESTS.append(
         "Rotated 3d",
         "3D",
         1e-1,
-        Rotated(KEYS, [random.uniform(np.pi / 6, np.pi) for _ in range(3)], True),  # type: ignore
+        Rotated(KEYS, [random.uniform(np.pi / 6, np.pi) for _ in range(3)], True, dtype=np.float64),  # type: ignore
     )
 )
 
@@ -229,7 +234,7 @@ TESTS.append(
         "RandRotated 3d",
         "3D",
         1e-1,
-        RandRotated(KEYS, *[random.uniform(np.pi / 6, np.pi) for _ in range(3)], 1),  # type: ignore
+        RandRotated(KEYS, *[random.uniform(np.pi / 6, np.pi) for _ in range(3)], 1, dtype=np.float64),  # type: ignore
     )
 )
